기본 리스트의 userid를 member의 id를 참조하는 외래키로 재생성, 이에 따른 전반적인 수정작업. 추가로 리스트를 작성하지 않은 제 3자의 수정, 삭제 제한및 경고 알림 추가

+++ src/main/java/com/example/database_project/controller/ListController.java
@@ -1,6 +1,7 @@
 package com.example.database_project.controller;
 
 import com.example.database_project.dto.ListDTO;
+import com.example.database_project.dto.MemberDTO;
 import com.example.database_project.service.ListService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -10,6 +11,8 @@ import org.springframework.data.web.PageableDefault;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
+import jakarta.servlet.http.HttpSession;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import java.io.IOException;
 import java.util.List;
@@ -23,7 +26,7 @@ public class ListController {
 
     // 리스트 전체 조회
     @GetMapping
-    public String list(@PageableDefault(size = 10, sort = "list_created_time", direction = Sort.Direction.DESC) Pageable pageable, Model model) {
+    public String list(@PageableDefault(size = 10, sort = "list_created_time", direction = Sort.Direction.DESC) Pageable pageable, Model model, HttpSession session) {
         Page<ListDTO> listPage = listService.paging(pageable);
         model.addAttribute("listPage", listPage);
         return "list"; // 반환하는 뷰 이름
@@ -32,37 +35,55 @@ public class ListController {
 
     // 저장 폼 요청
     @GetMapping("/save")
-    public String saveForm() {
+    public String saveForm(HttpSession session) {
+        System.out.println("session result: " + session.getAttribute("id"));
+        System.out.println("session result: " + session.getAttribute("memberEmail"));
+        System.out.println("session result: " + session.getAttribute("memberPassword"));
         return "Lsave";
     }
 
     // 데이터 저장
     @PostMapping("/save")
-    public String save(@ModelAttribute ListDTO listDTO) throws IOException {
-        listService.save(listDTO);
+    public String save(@ModelAttribute ListDTO listDTO, HttpSession session) {
+        listService.save(listDTO, session);
+        System.out.println("session result: " + session.getAttribute("id"));
         return "redirect:/list";
     }
 
     // 업데이트 폼 요청
     @GetMapping("/update/{id}")
-    public String updateForm(@PathVariable long id, Model model) {
+    public String updateForm(@PathVariable long id, Model model, HttpSession session, RedirectAttributes redirectAttributes) {
 
         ListDTO listDTO = listService.findBylID(id);
         model.addAttribute("list", listDTO);
-        return "Lupdate";
+        if((long) session.getAttribute("id") != listDTO.getMemberId()) {
+            redirectAttributes.addFlashAttribute("alertMessage", "You are not authorized to delete this item!");
+            return "redirect:/list";
+        }
+        else{
+            return "Lupdate";
+        }
+
     }
 
     // 데이터 업데이트
     @PostMapping("/update")
-    public String update(@ModelAttribute ListDTO listDTO) {
-        listService.update(listDTO);
+    public String update(@ModelAttribute ListDTO listDTO, HttpSession session) {
+        listService.update(listDTO, session);
         return "redirect:/list";
     }
 
     // 데이터 삭제
     @GetMapping("/delete/{id}")
-    public String delete(@PathVariable long id) {
-        listService.deleteById(id);
+    public String delete(@PathVariable long id, HttpSession session, Model model, RedirectAttributes redirectAttributes) {
+        ListDTO listDTO = listService.findBylID(id);
+        if((long) session.getAttribute("id") == listDTO.getMemberId()) {
+            listService.deleteById(id);
+        }
+        else{
+            redirectAttributes.addFlashAttribute("alertMessage", "You are not authorized to delete this item!");
+
+        }
         return "redirect:/list";
     }
